Fixed conversion warnings found by clang.

+++ HeroesPath/main.cpp
@@ -60,7 +60,7 @@ int main()
         sfml_util::Display::LogAllFullScreenVideoModes();
         sfml_util::Display::LogAllSupportedFullScreenVideoModes();
         sfml_util::WinSPtr_t winSPtr{ sfml_util::Display::OpenRenderWindow("Heroes' Path", sf::Style::Fullscreen, 0/*default to antialiasing disabled*/) };
-        winSPtr->setFramerateLimit( gameDataFileSPtr->GetCopyInt("system-window-frame-rate-limit", 0) );
+        winSPtr->setFramerateLimit(static_cast<unsigned int>(gameDataFileSPtr->GetCopyInt("system-window-frame-rate-limit", 0)) );
         winSPtr->setVerticalSyncEnabled( gameDataFileSPtr->GetCopyBool("system-window-sync", true) );
 
         //set resource paths for manager classes
